image-status: use supported=True by default to limit releases.
Previously we would build a list of releases to limit by:
  release~(list|of|supported|releases)
where the list came from 'distro-info --supported'.

That relied on distro-info getting updated.  In a apt installation,
that *would* get updated, but in a snap install it would not.  The result
being that we'd have a static list of "supported" releases until the
snap was re-built.

Instead now we use 'supported=True' and rely upon the server to tell
us what is supported and what is not.