Adjust how supported is done and fix ubuntu.
Previous commit adding  maas image status broke the supported filter
that was occuring in Ubuntu.  The change here is to use 'supported=True'
whenever we can (for cloud-*, ec2*) and use the supported information
from distro_info when we cannot.  By using supported=True we reduce
the pain of the fact that distro_info gets out dated in the snap.